fix(auth): use params instead of searchParams for callbacks
