+++ apps/editor/messages/pt.po
@@ -347,27 +347,22 @@ msgctxt "lP29gk"
 msgid "A course with this URL already exists"
 msgstr "Já existe um curso com este endereço"
 
-#: src/app/auth/callback/callback-handler.tsx
+#: src/app/auth/callback/[token]/page.tsx
 msgctxt "57VhQb"
 msgid "Authentication Error"
 msgstr "Erro de autenticação"
 
-#: src/app/auth/callback/callback-handler.tsx
-msgctxt "ARSMi2"
-msgid "The authentication token is missing. Please try logging in again."
-msgstr "O token de autenticação está faltando. Por favor, tente fazer login novamente."
-
-#: src/app/auth/callback/callback-handler.tsx
+#: src/app/auth/callback/[token]/page.tsx
 msgctxt "Le3MTe"
 msgid "Return to login"
 msgstr "Voltar para o login"
 
-#: src/app/auth/callback/callback-handler.tsx
+#: src/app/auth/callback/[token]/page.tsx
 msgctxt "mffe4M"
 msgid "The authentication token is invalid or expired. Please try logging in again."
 msgstr "O token de autenticação é inválido ou expirou. Por favor, tente fazer login novamente."
 
-#: src/app/auth/callback/callback-handler.tsx
+#: src/app/auth/callback/[token]/page.tsx
 msgctxt "vQjckV"
 msgid "Need help? Contact us at hello@zoonk.com"
 msgstr "Precisa de ajuda? Entre em contato com a gente em contato@zoonk.com"
