+++ apps/editor/src/app/auth/callback/[token]/page.tsx
@@ -13,12 +13,15 @@ import {
 import { FullPageLoading } from "@zoonk/ui/components/loading";
 import { AlertCircleIcon } from "lucide-react";
 import Link from "next/link";
-import { redirect, useSearchParams } from "next/navigation";
-import { Suspense, useEffect, useEffectEvent, useState } from "react";
+import { redirect } from "next/navigation";
+import { useExtracted } from "next-intl";
+import { use, useEffect, useEffectEvent, useState } from "react";
 
-type CallbackErrorType = "missing" | "invalid";
+type CallbackErrorType = "invalid";
 
 function CallbackError({ type }: { type: CallbackErrorType }) {
+  const t = useExtracted();
+
   return (
     <div className="flex min-h-dvh items-center justify-center p-4">
       <Empty className="border-none">
@@ -26,11 +29,12 @@ function CallbackError({ type }: { type: CallbackErrorType }) {
           <EmptyMedia variant="icon">
             <AlertCircleIcon aria-hidden="true" />
           </EmptyMedia>
-          <EmptyTitle>Authentication Error</EmptyTitle>
+          <EmptyTitle>{t("Authentication Error")}</EmptyTitle>
           <EmptyDescription>
-            {type === "missing"
-              ? "The authentication token is missing. Please try logging in again."
-              : "The authentication token is invalid or expired. Please try logging in again."}
+            {type === "invalid" &&
+              t(
+                "The authentication token is invalid or expired. Please try logging in again.",
+              )}
           </EmptyDescription>
         </EmptyHeader>
 
@@ -39,10 +43,10 @@ function CallbackError({ type }: { type: CallbackErrorType }) {
             className={buttonVariants({ variant: "outline" })}
             href="/login"
           >
-            Return to login
+            {t("Return to login")}
           </Link>
           <p className="text-muted-foreground text-sm">
-            Need help? Contact us at hello@zoonk.com
+            {t("Need help? Contact us at hello@zoonk.com")}
           </p>
         </EmptyContent>
       </Empty>
@@ -50,9 +54,10 @@ function CallbackError({ type }: { type: CallbackErrorType }) {
   );
 }
 
-function CallbackHandler() {
-  const searchParams = useSearchParams();
-  const token = searchParams.get("token");
+export default function AuthCallbackPage(
+  props: PageProps<"/auth/callback/[token]">,
+) {
+  const { token } = use(props.params);
   const [error, setError] = useState<CallbackErrorType | null>(null);
 
   const handleVerify = useEffectEvent(async (userToken: string) => {
@@ -70,12 +75,7 @@ function CallbackHandler() {
   });
 
   useEffect(() => {
-    if (token) {
-      void handleVerify(token);
-    } else {
-      console.info("Auth callback: missing token");
-      setError("missing");
-    }
+    void handleVerify(token);
   }, [token]);
 
   if (error) {
@@ -83,12 +83,4 @@ function CallbackHandler() {
   }
 
   return <FullPageLoading />;
-}
-
-export default function AuthCallbackPage() {
-  return (
-    <Suspense fallback={<FullPageLoading />}>
-      <CallbackHandler />
-    </Suspense>
-  );
 }
