+++ apps/main/messages/en.po
@@ -505,27 +505,22 @@ msgctxt "YdQxfw"
 msgid "Contact Support"
 msgstr "Contact Support"
 
-#: src/app/[locale]/auth/callback/page.tsx
+#: src/app/[locale]/auth/callback/[token]/page.tsx
 msgctxt "57VhQb"
 msgid "Authentication Error"
 msgstr "Authentication Error"
 
-#: src/app/[locale]/auth/callback/page.tsx
-msgctxt "ARSMi2"
-msgid "The authentication token is missing. Please try logging in again."
-msgstr "The authentication token is missing. Please try logging in again."
-
-#: src/app/[locale]/auth/callback/page.tsx
+#: src/app/[locale]/auth/callback/[token]/page.tsx
 msgctxt "Le3MTe"
 msgid "Return to login"
 msgstr "Return to login"
 
-#: src/app/[locale]/auth/callback/page.tsx
+#: src/app/[locale]/auth/callback/[token]/page.tsx
 msgctxt "mffe4M"
 msgid "The authentication token is invalid or expired. Please try logging in again."
 msgstr "The authentication token is invalid or expired. Please try logging in again."
 
-#: src/app/[locale]/auth/callback/page.tsx
+#: src/app/[locale]/auth/callback/[token]/page.tsx
 msgctxt "vQjckV"
 msgid "Need help? Contact us at hello@zoonk.com"
 msgstr "Need help? Contact us at hello@zoonk.com"
