+++ apps/main/src/app/[locale]/auth/callback/[token]/loading.tsx
@@ -1,0 +1,5 @@
+import { FullPageLoading } from "@zoonk/ui/components/loading";
+
+export default function Loading() {
+  return <FullPageLoading />;
+}
